+++ classes/matchHandler.js
@@ -3,11 +3,10 @@ const { MATCH_SYMBOL } = require("../constants/constants.js");
 
 class MatchHandler {
   /* --------------------------------
-  * MAIN
-  * -------------------------------- */
-  static clearMatches(grid) {
+   * MAIN
+   * -------------------------------- */
+  static handleMatches(grid) {
     let columns = MatchHandler.getColumnsFromGrid(grid);
-
     for (let col = 0; col < columns.length; col++) {
       let column = columns[col];
       let colAfterFall = MatchHandler.makeAllGemsFall(column);
@@ -21,6 +20,11 @@ class MatchHandler {
   /* --------------------------------
    * MAKE GEMS FALL
    * -------------------------------- */
+  /**
+   * makeAllGemsFall
+   * @param {*} column - a column of {@link Gem} instances 
+   * @returns column of gems after gems have fallen
+   */
   static makeAllGemsFall(column) {
     let lowestStar = MatchHandler.findLowestStar(column);
     let lowestGemAboveStar = MatchHandler.findLowestGemAboveStar(column, lowestStar);
@@ -39,7 +43,8 @@ class MatchHandler {
   }
 
   /**
-   * Finds lowest star in a column (i.e. the lowest gem in a match)
+   * findLowestStar
+   * finds lowest star in a column (i.e. the lowest gem in a match)
    * This helps determine how far the gems should fall
    * @param {*} column - a column of {@link Gem} instances 
    * @returns {void} 
@@ -54,7 +59,8 @@ class MatchHandler {
   }
 
   /**
-   * Finds lowest non-matched gem above a particular star (i.e. matched gem)
+   * findLowestGemAboveStar
+   * finds lowest non-matched gem above a particular star (i.e. matched gem)
    * This helps determine where the gems should start falling 
    * @param {*} column - a column of {@link Gem} instances  
    * @param {*} star - a {@link Gem} instance that is part of a match
@@ -83,7 +89,7 @@ class MatchHandler {
   }
 
   /* --------------------------------
-   * SWITCH BETWEEN COLS & ROWS
+   * SWITCH BETWEEN COLS & ROWS (i.e. GRID)
    * -------------------------------- */
   static getGridFromColumns = (columns) => this.transpose(columns);
   static getColumnsFromGrid = (rows) => this.transpose(rows);
